+++ packages/@weresk/core/src/utils/objects/objectMap.ts
@@ -1,6 +1,3 @@
-// eslint-disable-next-line
-export function objectMap(obj: object, fn: Function) {
-    return Object.fromEntries(
-        Object.entries(obj).map(([k, v], i) => [k, fn(v, k, i)])
-    );
+export function objectMap(obj: object, fn: (k: string, v: any, i: number) => any): object {
+    return Object.fromEntries(Object.entries(obj).map(([k, v], i) => [k, fn(k, v, i)]));
 }